moved customerstatus enum
