+++ src/enums/customerStatus.ts
@@ -1,0 +1,7 @@
+enum CustomerStatus {
+  ForReview = "For Review",
+  Accepted = "Accepted",
+  ChangeRequested = "Change Requested",
+}
+
+export default CustomerStatus
